check if digit (0-9)

+++ 0x04-more_functions_nested_loops/1-isdigit.c
@@ -13,7 +13,7 @@ int _isdigit(int c)
 
 	for (a = '0'; a <= '9'; a++)
 	{
-		if (c == a)
+		if ((int) c == a)
 			b = 1;
 	}
 	return (b);
